// web/convex/_generated/api.d.ts
declare const fullApi: ApiFromModules<{
  _helpers: typeof _helpers;
  channels: typeof channels;
  http: typeof http;
  invites: typeof invites;
  members: typeof members;
  messages: typeof messages;
  users: typeof users;
  workspaces: typeof workspaces;
}>;

/**
 * A utility for referencing Convex functions in your app's public API.
 *
 * Usage:
 * ```js
 * const myFunctionReference = api.myModule.myFunction;
 * ```
 */
export declare const api: FilterApi<
  typeof fullApi,
  FunctionReference<any, "public">
>;